fast test compilation creating adoc api util
