+++ src/prover_server.cc
@@ -19,6 +19,7 @@
 #include "zethConfig.h"
 
 #include "util.hpp"
+#include "util_api.hpp"
 
 // Include the file generated by gRPC
 #include "prover.grpc.pb.h"
